Javascript+buttons+footer added
I added some Javascript to the file so that when you click on certain titles on the sidebar, the information is then displayed in the main section of the guidebook. I also added buttons to the other areas of the website (the prompt generator and the prompt browser) and then I added a footer as well.