Pagination for Prompts
I got it to where the prompts will all load on the "prompt browser" page and after 6 have populated on one page, they will go on to the next page (page 2). I also updated some of the CSS/styling of the page. Next task is to filter the prompts based on certain parameters.